Fix compilation on x86_64
git-svn-id: file:///openmsx-svn/code/openmsx/trunk@13140 7512d703-071c-0410-b5c2-d2bdd895b7d1

+++ src/video/BitmapConverter.cc
@@ -159,7 +159,7 @@ void BitmapConverter<Pixel>::renderGraphic4(
 	}
 #endif
 
-	if ((sizeof(Pixel) == 2) && (((int)pixelPtr & 1) == 1)) {
+	if ((sizeof(Pixel) == 2) && ((long(pixelPtr) & 1) == 1)) {
 		// Its 16 bit destination but currently not aligned on a word boundary
 		// First write one pixel to get aligned
 		// Then write double pixels in a loop with 4 double pixels (is 8 single pixels) per iteration
